Create get-ROI-timeseries.ijm
Macro to get time series of grid-like regions of interest across an image stack. Requires at least 1 open image stack, which can be multichannel, and at least 1 ROI in the ROI Manager - this will be the area to overlay the grid onto to extract time series. You can have different ROIs for different image stacks, however the macro assumes the ROIs are all the same size, just in different locations across the different image stacks. 

A GUI will also for the file path location to save the time series and associated information to. The GUI will also ask you to pair the image stacks and an ROI, so that you can choose which ROI to use for which image stack. 

The macro will return .csv files with ROI time series for each channel of each image stack, as well as a meta-file with dimensions of the image stacks.